MVP for credentials management
This commit touches on many things:

+ Introduction of a UserCredentialsView that handles the input of the user, password and server for your Nextcloud instance

The user+password get saved to the keychain with the server as key (used for retrieval later)

The server is not sensitive so we store it to UserDefaults

We also use some (probably too many) booleans to handle some UIState like hasCredentials / hasCredentialsDrilled

+ KeychainManager module, Credentials and KeychainError models have been added for that matter

+ There's some dependency injection of Network so that the whole app uses the same instance that gets the credentials from MainView at load time.

This might change now that I saw how to just inject the specific methods (refreshBookmarks is the example) I need rather than the whole Network object.

The approach mentioned above will also make testing easier.

Next steps include sharing credentials with the share extension so that we don't have them hardcoded there either.